Updated python and matlab code to cache values of profile, peak objects
Now, matrices A, L, etc are calculated once during initialization, which speeds up REGALS iterations. This sacrfices the ability to edit the parameters (such as dmax) without reinitializing the concentration, profle objects. However, none of the demos rely on the ability to edit parameters, so in practice this is not a problem.

In MATLAB: This was done in Concentration.m and Profile.m. The properties are read-only, and the regularizer object is now private (requiring small changes to the demo scripts).

In Python: Was done in concentration_class and profile_class. The properties are not read-only (would have required a set/get interface), so it is up to the user not to modify them.